add to session test to validate not all sessions are deleted
it would be much nicer to update DateTimeFreezer and use that - but
the JDBI code does not use JodaTime, therefore it doesn't work.  The
only way to do it now would be to use a Clock in SessionRepositoryBase
to get the time and update it for tests, however the changes are
invasive and would allow people to change the clock in the repository
in `main` code.  Plus I couldn't get it to work properly.

This tests just does a Thread.sleep() :(